requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.8
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
